Robot_Pokja/requirements.txt
google-generativeai
pdfplumber
python-dotenv
pandas
openpyxl
selenium
webdriver-manager
